[FIX] orgnasied directory structure
