+++ src/io/keyboard.rs
@@ -1,4 +1,60 @@
-pub static TO_ASCII: [Option<char>; 256] = {
+use crate::include::asm_utile;
+use spin::Mutex;
+
+const KEYBOARD_STATUS_PORT: u16 = 0x64;
+const KEYBOARD_DATA_PORT: u16 = 0x60;
+const SHIFT_LEFT: u8 = 0x2A;
+const SHIFT_RIGHT: u8 = 0x36;
+const SHIFT_LEFT_RELEASE: u8 = 0x2A + 0x80;
+const SHIFT_RIGHT_RELEASE: u8 = 0x36 + 0x80;
+
+static SHIFT_PRESSED: Mutex<bool> = Mutex::new(false);
+
+pub fn init() {
+	unsafe {
+		asm_utile::outb(KEYBOARD_STATUS_PORT, 0xAE);
+
+		while asm_utile::inb(KEYBOARD_STATUS_PORT) & 0x01 != 0 {
+			let _ = asm_utile::inb(KEYBOARD_DATA_PORT);
+		}
+
+		asm_utile::outb(KEYBOARD_DATA_PORT, 0xF4);
+	}
+}
+
+pub fn read() -> Option<char> {
+	let scancode: u8;
+	let key: Option<char>;
+	let mut shift: spin::MutexGuard<'_, bool> = SHIFT_PRESSED.lock();
+
+	unsafe {
+		scancode = asm_utile::inb(KEYBOARD_DATA_PORT);
+	}
+
+	match scancode {
+		SHIFT_LEFT | SHIFT_RIGHT => {
+			*shift = true;
+			return None;
+		}
+		SHIFT_LEFT_RELEASE | SHIFT_RIGHT_RELEASE => {
+			*shift = false;
+			return None;
+		}
+		_ => {
+			if scancode & 0x80 == 0 {
+				key = if *shift {
+					TO_SHIFT_ASCII[scancode as usize]
+				} else {
+					TO_ASCII[scancode as usize]
+				};
+				return key;
+			}
+		}
+	}
+	None
+}
+
+static TO_ASCII: [Option<char>; 256] = {
 	let mut table = [None; 256];
 
 	table[0x02] = Some('1');
@@ -53,7 +109,7 @@ pub static TO_ASCII: [Option<char>; 256] = {
 	table
 };
 
-pub static TO_SHIFT_ASCII: [Option<char>; 256] = {
+static TO_SHIFT_ASCII: [Option<char>; 256] = {
 	let mut table = [None; 256];
 
 	table[0x02] = Some('!');
