+++ src/io/mod.rs
@@ -1,0 +1,3 @@
+pub mod keyboard;
+pub mod shell;
+pub mod vga_buffer;
