--- src/keyboard/mod.rs
@@ -1,3 +1,0 @@
-pub mod keyboard;
-pub use keyboard::*;
-pub mod keymap;
