--- src/shell/mod.rs
@@ -1,2 +1,0 @@
-pub mod shell;
-pub use shell::*;
